Refactored : Added RF2 Generic Class
